rentalsite-main/requirements.txt
asgiref==3.7.2
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.1
cryptography==41.0.5
defusedxml==0.7.1
Django==4.2.6
django-allauth==0.58.2
django-autocomplete-light==3.9.7
django-autoslug==1.9.9
django-crispy-forms==2.1
djangorestframework==3.14.0
idna==3.4
oauthlib==3.2.2
Pillow==10.1.0
progressbar2==4.2.0
psycopg2==2.9.9
pycparser==2.21
PyJWT==2.8.0
python-social-auth==0.3.6
python-utils==3.8.1
python3-openid==3.2.0
pytz==2023.3.post1
requests==2.31.0
requests-oauthlib==1.3.1
six==1.16.0
social-auth-app-django==5.4.0
social-auth-core==4.5.0
sqlparse==0.4.4
typing_extensions==4.8.0
tzdata==2023.3
Unidecode==1.3.7
urllib3==2.0.7
